+++ lib/src/b1enumtypes.dart
@@ -1,9 +1,13 @@
-const tNO = "tNO";
-const tYES = "tYES";
-//            <EnumType Name="BoActivities">
-const cn_Conversation = "cn_Conversation";
-const cn_Meeting = "cn_Meeting";
-const cn_Task = "cn_Task";
-const cn_Other = "cn_Other";
-const cn_Note = "cn_Note";
-const cn_Campaign = "cn_Campaign";
+///A base class sor B1 enumerations
+abstract class BoEnums {
+  final String value;
+  const BoEnums(this.value) : assert(value != null);
+  @override bool operator == (v) => this.value == v.toString();
+  @override String toString() => value;
+}
+
+class BoYesNoEnum extends BoEnums {
+  static const tNO = BoYesNoEnum("tNO");
+  static const tYES = BoYesNoEnum("tYES");
+  const BoYesNoEnum(String value) : super(value);
+}